+++ store/storetest/kvstoretest.go
@@ -3,6 +3,7 @@ package storetest
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/dfuse-io/kvdb/store"
@@ -82,6 +83,18 @@ func TestBasic(t *testing.T, driver store.KVStore) {
 	testPrefix(t, driver, []byte("b"), 3, all[1:4])
 	testPrefix(t, driver, []byte("ba"), 10, all[1:4])
 
+	// testing BatchPrefix without limit
+	testBatchPrefix(t, driver, [][]byte{[]byte("ba")}, store.Unlimited, all[1:4]...)
+	testBatchPrefix(t, driver, [][]byte{[]byte("ba"), []byte("c")}, store.Unlimited, all[1], all[2], all[3], all[5])
+	testBatchPrefix(t, driver, [][]byte{[]byte("a"), []byte("c")}, store.Unlimited, all[0], all[5])
+	testBatchPrefix(t, driver, [][]byte{[]byte("d"), []byte("f")}, store.Unlimited)
+
+	// testing BatchPrefix with limit
+	testBatchPrefix(t, driver, [][]byte{[]byte("ba")}, 1, all[1])
+	testBatchPrefix(t, driver, [][]byte{[]byte("ba"), []byte("c")}, 2, all[1], all[2])
+	testBatchPrefix(t, driver, [][]byte{[]byte("a"), []byte("c")}, 1, all[0])
+	testBatchPrefix(t, driver, [][]byte{[]byte("d"), []byte("f")}, 10)
+
 	// testing Scan without limit
 	testScan(t, driver, []byte("a"), []byte("a"), store.Unlimited, nil)
 	testScan(t, driver, []byte("a"), []byte("b"), store.Unlimited, all[:1])
@@ -126,6 +139,23 @@ func testPrefix(t *testing.T, driver store.KVStore, prefix []byte, limit int, ex
 	require.Equal(t, exp, got)
 }
 
+func testBatchPrefix(t *testing.T, driver store.KVStore, prefixes [][]byte, limit int, exp ...store.KV) {
+	var got []store.KV
+	itr := driver.BatchPrefix(context.Background(), prefixes, limit)
+	for itr.Next() {
+		got = append(got, itr.Item())
+	}
+
+	stringPrefixes := make([]string, len(prefixes))
+	for i, prefix := range prefixes {
+		stringPrefixes[i] = string(prefix)
+	}
+
+	testPrintKVs(fmt.Sprintf("test prefixes with prefix %q", strings.Join(stringPrefixes, ", ")), got)
+	require.NoError(t, itr.Err())
+	require.Equal(t, exp, got)
+}
+
 func testScan(t *testing.T, driver store.KVStore, start, end []byte, limit int, exp []store.KV) {
 	var got []store.KV
 	itr := driver.Scan(context.Background(), start, end, limit)
@@ -146,8 +176,10 @@ func testStringsToKey(parts ...string) (out []byte) {
 }
 
 func testPrintKVs(title string, out []store.KV) {
-	//fmt.Printf("%s\n", title)
-	//for _, kv := range out {
-	//	fmt.Printf("- %s => %s\n", string(kv.Key), string(kv.Value))
-	//}
+	if debug {
+		fmt.Printf("%s\n", title)
+		for _, kv := range out {
+			fmt.Printf("- %s => %s\n", string(kv.Key), string(kv.Value))
+		}
+	}
 }