+++ store/storetest/testing.go
@@ -1,10 +1,17 @@
 package storetest
 
 import (
+	"os"
 	"testing"
 
 	"github.com/dfuse-io/kvdb/store"
 )
+
+var debug = false
+
+func init() {
+	debug = os.Getenv("DEBUG") != ""
+}
 
 type DriverCleanupFunc func()
 type DriverFactory func() (store.KVStore, DriverCleanupFunc)