+++ store/tikv/tikv_test.go
@@ -1,6 +1,7 @@
 package tikv
 
 import (
+	"fmt"
 	"io"
 	"os"
 	"testing"
@@ -17,16 +18,19 @@ func init() {
 }
 
 func TestAll(t *testing.T) {
-	if os.Getenv("TEST_TIKV") != "" {
-		storetest.TestAll(t, "tikv", newTestFactory(t))
+	if os.Getenv("TEST_TIKV") == "" {
+		t.Skip("To run those tests, you need to have TEST_TIKV environment variable set")
+		return
 	}
+
+	storetest.TestAll(t, "tikv", newTestFactory(t))
 }
 
 func newTestFactory(t *testing.T) storetest.DriverFactory {
 	return func() (store.KVStore, storetest.DriverCleanupFunc) {
-		kvStore, err := NewStore("tikv://pd0:2379/helloworld")
+		kvStore, err := NewStore("tikv://localhost:2379/data")
 		if err != nil {
-			t.Skip("pd0 unreachable, cannot run tests") // FIXME: this just times out
+			t.Skip(fmt.Errorf("pd0 unreachable, cannot run tests: %w", err)) // FIXME: this just times out
 			return nil, nil
 		}
 		require.NoError(t, err)